+++ packages/client/src/constants.ts
@@ -26,7 +26,8 @@ export const createConstantEntry =
       combineLatest([chainHead.getRuntimeContext$(at || null), getCompat]).pipe(
         map(([ctx, compat]) => {
           const value = constFromCtx(ctx, pallet, name)
-          if (compat(ctx).isValueCompatible(value)) return value
+          if (compat(ctx).getValueCompatibility(value).type === "compatible")
+            return value
           throw new IncompatibleRuntimeError("Constant", `${pallet}.${name})`)
         }),
       ),