+++ packages/client/src/utils/stg-get-key.ts
@@ -1,22 +1,36 @@
-import { IncompatibleRuntimeError, InvalidArgsError } from "@/compatibility"
+import {
+  IncompatibleRuntimeError,
+  InvalidArgsError,
+  ValueCompatibility,
+} from "@/compatibility"
 import { RuntimeContext } from "@polkadot-api/observable-client"
 
 export const stgGetKey = (
   ctx: RuntimeContext,
   pallet: string,
   name: string,
-  areAgsCompat: (args: Array<any>) => boolean,
+  getArgsCompatibility: (args: Array<any>) => ValueCompatibility,
 ) => {
   let codecs
   try {
     codecs = ctx.dynamicBuilder.buildStorage(pallet, name)
   } catch {
-    throw new IncompatibleRuntimeError("Storage", `${pallet}.${name}`)
+    throw new Error(`Runtime entry Storage(${pallet}.${name}) not found`)
   }
 
   return (...args: any[]) => {
-    if (args.length === codecs.len && !areAgsCompat(args))
-      throw new InvalidArgsError("Storage", `${pallet}.${name}`, args)
+    if (args.length === codecs.len) {
+      const compatibilityResult = getArgsCompatibility(args)
+      if (compatibilityResult.type === "incompatible")
+        throw new InvalidArgsError(
+          "Storage",
+          `${pallet}.${name}`,
+          args,
+          compatibilityResult.value,
+        )
+      if (compatibilityResult.type === "runtimeIncompatible")
+        throw new IncompatibleRuntimeError("Storage", `${pallet}.${name}`)
+    }
     return codecs.keys.enc(...args)
   }
 }